Rewrite triangle randomizer to spawn triangles more correctly
This change should fix the case where almost all triangles spawn outside of the circle or close to the edges. Triangle values are also mapped to the limits, which will reduce the number of identical values. The code that was responsible for moving the triangles when passing a distance equal to their size was removed, as this caused random teleportation of the triangles during the explosion.